Revert "Move name_repr, index_repr into common.formatting"
This reverts commit f1a97393102e2e461be944a02f829eae16206ad8.
